requirements.txt
Django==1.2.3
Markdown==2.0.3
MySQL-python==1.2.3
South==0.7.2
distribute==0.6.14
django-email-confirmation==0.1.4
django-extensions==0.5
django-markitup==0.6.1
django-robots==0.8.0
django-uni-form==0.7.0
docutils==0.7
facebook-python-sdk==0.1
httplib2==0.6.0
oauth2==1.2.0
python-openid==2.2.5
wsgiref==0.1.2
